requirements.txt
Flask
scikit-learn
SQLAlchemy
python-dotenv
joblib
pandas
numpy
psycopg2-binary